requirements.txt
pytest==9.0.1
pytest-cov==5.0.0
pytest-xdist==3.5.0
pytest-timeout==2.2.0
pytest-allure==1.1.1.post1
requests==2.31.0
pyyaml==6.0.1
python-dotenv==1.0.0
tenacity==8.2.3
pymysql==1.1.0
allure-pytest==2.13.2
allure-behave==2.13.2
